+++ js/api/account/login-account.mjs
@@ -1,4 +1,4 @@
-import { authenticateProfile } from "../auth/authenticate-profile.mjs";
+import { authenticateProfile } from "../authenticate/authenticate-profile.mjs";
 
 export function loginAccount(event) {
    const formElements = event.target;
